requirements.txt
# Django and REST framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0

# Database
mysqlclient==2.2.0

# Authentication
django-allauth==0.57.0

# API utilities
django-filter==23.5
drf-yasg==1.21.7

# Environment variables
python-decouple==3.8

# Image processing
Pillow==10.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
